login logout db skeloton code
