+++ README.md
@@ -15,6 +15,12 @@ Please speak to @Thundersnail before changing any database stuff.
 ##instruction
 from terminal in 1 directory up run: 
 python3
->>> export FLASK_APP=appName
->>> export FLASK_DEBUG=1
->>> flask run
+>>>from appName import db, create_app
+>>>db.create_all(app=create_app)
+
+run commands:
+export FLASK_APP=appName
+#doesnt work if ap name has '-'
+#changed it on my computer to '_'
+export FLASK_DEBUG=1
+flask run
