+++ server.py
@@ -1,26 +1,44 @@
-import flask
-import config
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
+from flask_restful import Resource, Api
+db = SQLAlchemy()
 
-import db
+def create_app():
+    app = Flask(__name__)
+    api = Api(app)
 
-app = flask.Flask(__name__)
+    app.config['SECRET_KEY']='thisismysecretkey'
+    app.config['SQLALCHEMY_DATABASE_URI']='sqlite:///db.sqlite3'
 
+    db.init_app(app)
 
-@app.route('/')
-def hello_world():
-    return flask.render_template("index.html", card_info="Display video here.")
+    login_manager =  LoginManager()
+    login_manager.login_view = 'pages.login'
+    login_manager.init_app(app)
+
+    from models import User
+    @login_manager.user_loader
+    def load_user(user_id):
+        return User.query.get(int(user_id))
+
+    from pages import pages as pagesBlueprint
+    app.register_blueprint(pagesBlueprint)
+
+    return app
 
 
 if __name__ == '__main__':
     def main():
-        # app.run(debug=config.debug)
-        app.run(debug=False)
-
-    def db_test_main():
-        session = db.SqlAlchemySession()
-        user = session.query(db.User).filter_by(id=1).first()
-        session.commit()
-        print(repr(user))
+        app = create_app()
+        app.run(debug=config.debug)
+        app.run(debug=True)
 
     main()
-    # db_test_main()
+
+#     def db_test_main():
+#         session = db.SqlAlchemySession()
+#         user = session.query(db.User).filter_by(id=1).first()
+#         session.commit()
+#         print(repr(user))
+#     db_test_main()
