+++ templates/index.html
@@ -4,9 +4,9 @@
 
 {% block body %}
 <h1 class="title">
-    Flask Login Example
+    HOME PAGE
 </h1>
 <h2 class="subtitle">
-    Easy authentication and authorization in Flask.
+    HOME PAGE FOR {{ current_user.name }}
 </h2>
 {% endblock %}
